* edit the Protocol

+++ src/main/java/jndc/client/JNDCClientMessageHandle.java
@@ -95,7 +95,8 @@ public class JNDCClientMessageHandle extends SimpleChannelInboundHandler<NDCMess
 
     @Override
     protected void channelRead0(ChannelHandlerContext channelHandlerContext, NDCMessageProtocol ndcMessageProtocol) throws Exception {
-        Integer type = ndcMessageProtocol.getType();
+        byte type = ndcMessageProtocol.getType();
+
         try {
